pyproject.toml
[project]
name = "oos-workflow"
version = "0.1.0"
description = "Simple Python workflow package for Azure Databricks testing"
requires-python = ">=3.11"
dependencies = [
    "pandas==2.2.0",
    "pyspark==3.5.0",
    "pyyaml==6.0.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
